Relocate utility functions to complete TODO
