+++ cstar/cli/workplan/plan.py
@@ -6,11 +6,11 @@ import matplotlib.pyplot as plt
 import networkx as nx
 import typer
 
+from cstar.base.utils import slugify
 from cstar.orchestration.models import Workplan
 from cstar.orchestration.orchestration import Planner
 from cstar.orchestration.serialization import deserialize
 from cstar.orchestration.transforms import RomsMarblTimeSplitter, WorkplanTransformer
-from cstar.orchestration.utils import slugify
 
 app = typer.Typer()
