+++ cstar/orchestration/utils.py
@@ -1,5 +1,4 @@
 import os
-import re
 import shutil
 import typing as t
 from pathlib import Path
@@ -18,25 +17,6 @@ ENV_CSTAR_SLURM_ACCOUNT: t.Literal["CSTAR_SLURM_ACCOUNT"] = "CSTAR_SLURM_ACCOUNT
 ENV_CSTAR_SLURM_QUEUE: t.Literal["CSTAR_SLURM_QUEUE"] = "CSTAR_SLURM_QUEUE"
 
 
-def slugify(source: str) -> str:
-    """Convert a source string into a URL-safe slug.
-
-    Parameters
-    ----------
-    source : str
-        The string to be converted.
-
-    Returns
-    -------
-    str
-        The slugified version of the source string.
-    """
-    if not source:
-        raise ValueError
-
-    return re.sub(r"\W+", "-", source.casefold())
-
-
 def clear_working_dir(path: Path) -> None:
     """Clear specific paths under the working directory if CSTAR_CLOBBER_WORKING_DIR is set.
 
@@ -52,26 +32,3 @@ def clear_working_dir(path: Path) -> None:
         print(f"clearing {path}")
         shutil.rmtree(path / "input", ignore_errors=True)
         shutil.rmtree(path / "output", ignore_errors=True)
-
-
-def deep_merge(d1: dict[str, t.Any], d2: dict[str, t.Any]) -> dict[str, t.Any]:
-    """Deep merge two dictionaries.
-
-    Parameters
-    ----------
-    d1 : dict[str, t.Any]
-        The first dictionary.
-    d2 : dict[str, t.Any]
-        The second dictionary.
-
-    Returns
-    -------
-    dict[str, t.Any]
-        The merged dictionaries.
-    """
-    for k, v in d2.items():
-        if isinstance(v, dict):
-            d1[k] = deep_merge(d1.get(k, {}), v)
-        else:
-            d1[k] = v
-    return d1
